test(usage): separate supported contracts from known limitations
