+++ tests/test_usage_accounting_audit.py
@@ -1,13 +1,18 @@
-"""Offline SDK contracts and reproductions for the September 2026 usage audit.
+"""Characterizations of known usage limitations, NOT desired permanent contracts.
 
-Tests named ``current`` characterize limitations, not a promised future API.
-They should change alongside an approved fix. No transports or mocks are used.
+When correcting a limitation, replace its negative assertion with a positive
+regression in test_usage_accounting_contracts.py (or the owning subsystem's
+suite), and update the audit finding. Do not preserve a bug to keep this file
+green, skip a newly fixed case, or rewrite its expectation to another bug.
+
+The audit report maps each case to its finding and replacement guidance.
+No transports or mocks are used. SDK payloads remain explicit in each case.
 """
 
 from __future__ import annotations
 
 import json
-from collections.abc import AsyncIterator
+from collections.abc import AsyncGenerator
 
 import pytest
 from openai.types import CompletionUsage
@@ -15,9 +20,9 @@ from openai.types.chat import ChatCompletion, ChatCompletionChunk
 from openai.types.responses import Response, ResponseUsage
 from pydantic import BaseModel
 
-from instructor.batch import BatchProcessor
+from instructor.batch import BatchProcessor, BatchSuccess
 from instructor.v2.core.providers import Provider
-from instructor.v2.core.retry import _initialize_usage, _usage_total_tokens
+from instructor.v2.core.retry import _initialize_usage
 from instructor.v2.core.usage import has_compatible_usage, update_total_usage
 from instructor.v2.providers.openai.handlers import OpenAIJSONHandler
 
@@ -33,7 +38,9 @@ def completion(usage: CompletionUsage) -> ChatCompletion:
     )
 
 
-def test_current_retry_usage_mutates_response_and_manufactures_details() -> None:
+def test_known_limitation_retry_usage_mutates_response_and_manufactures_details() -> (
+    None
+):
     response = completion(
         CompletionUsage(prompt_tokens=10, completion_tokens=5, total_tokens=15)
     )
@@ -41,7 +48,7 @@ def test_current_retry_usage_mutates_response_and_manufactures_details() -> None
     assert response.usage.prompt_tokens_details is None
     total = _initialize_usage(Provider.OPENAI)
     update_total_usage(response, total)
-    assert response.usage.prompt_tokens_details.cached_tokens == 0
+    assert response.usage.model_dump()["prompt_tokens_details"]["cached_tokens"] == 0
     second = completion(
         CompletionUsage(prompt_tokens=20, completion_tokens=7, total_tokens=27)
     )
@@ -51,7 +58,7 @@ def test_current_retry_usage_mutates_response_and_manufactures_details() -> None
     assert response.usage.total_tokens == 15
 
 
-def test_current_nested_dictionary_cost_is_last_attempt_only() -> None:
+def test_known_limitation_nested_dictionary_cost_is_last_attempt_only() -> None:
     total = _initialize_usage(Provider.OPENAI)
     for cost in [0.25, 0.75]:
         usage = CompletionUsage.model_validate(
@@ -68,96 +75,7 @@ def test_current_nested_dictionary_cost_is_last_attempt_only() -> None:
     assert total.model_extra["cost_details"]["upstream_inference_cost"] == 0.75
 
 
-def test_anthropic_thinking_detail_accumulates_across_sdk_shapes() -> None:
-    anthropic = pytest.importorskip("anthropic.types")
-    if "output_tokens_details" in anthropic.Usage.model_fields:
-        pytest.skip(
-            "Newer SDK declares this field; audit the installed shape separately"
-        )
-    total = _initialize_usage(Provider.ANTHROPIC)
-    for _ in range(2):
-        message = anthropic.Message.model_validate(
-            {
-                "id": "audit",
-                "type": "message",
-                "role": "assistant",
-                "content": [],
-                "model": "audit",
-                "stop_reason": "end_turn",
-                "stop_sequence": None,
-                "usage": {
-                    "input_tokens": 10,
-                    "output_tokens": 20,
-                    "output_tokens_details": {"thinking_tokens": 12},
-                },
-            }
-        )
-        update_total_usage(message, total)
-    assert total.output_tokens == 40
-    assert total.model_dump()["output_tokens_details"]["thinking_tokens"] == 24
-    assert message.usage.model_dump()["output_tokens_details"]["thinking_tokens"] == 24
-
-
-def test_openai_subsets_are_not_added_again_to_total() -> None:
-    usage = CompletionUsage.model_validate(
-        {
-            "prompt_tokens": 100,
-            "completion_tokens": 50,
-            "total_tokens": 150,
-            "prompt_tokens_details": {"cached_tokens": 80, "audio_tokens": 10},
-            "completion_tokens_details": {
-                "reasoning_tokens": 30,
-                "accepted_prediction_tokens": 5,
-                "rejected_prediction_tokens": 3,
-            },
-            "is_billable": True,
-        }
-    )
-    total = _initialize_usage(Provider.OPENAI)
-    for _ in range(2):
-        update_total_usage(completion(usage.model_copy(deep=True)), total)
-    assert _usage_total_tokens(total) == 300
-    assert total.prompt_tokens_details.cached_tokens == 160
-    assert total.completion_tokens_details.reasoning_tokens == 60
-    assert total.model_extra["is_billable"] is True
-
-
-def test_anthropic_cache_ttl_and_tools_have_separate_semantics() -> None:
-    anthropic = pytest.importorskip("anthropic.types")
-    total = _initialize_usage(Provider.ANTHROPIC)
-    for _ in range(2):
-        message = anthropic.Message.model_validate(
-            {
-                "id": "audit",
-                "type": "message",
-                "role": "assistant",
-                "content": [],
-                "model": "audit",
-                "stop_reason": "end_turn",
-                "stop_sequence": None,
-                "usage": {
-                    "input_tokens": 100,
-                    "output_tokens": 50,
-                    "cache_read_input_tokens": 200,
-                    "cache_creation_input_tokens": 30,
-                    "cache_creation": {
-                        "ephemeral_5m_input_tokens": 20,
-                        "ephemeral_1h_input_tokens": 10,
-                    },
-                    "server_tool_use": {
-                        "web_search_requests": 2,
-                        "web_fetch_requests": 1,
-                    },
-                },
-            }
-        )
-        update_total_usage(message, total)
-    assert _usage_total_tokens(total) == 760
-    assert total.cache_creation.ephemeral_1h_input_tokens == 20
-    assert total.server_tool_use.web_search_requests == 4
-
-
-def test_current_responses_api_usage_is_not_recognized() -> None:
+def test_known_limitation_responses_api_usage_is_not_recognized() -> None:
     response = Response.model_validate(
         {
             "id": "resp_audit",
@@ -185,7 +103,7 @@ def test_current_responses_api_usage_is_not_recognized() -> None:
     assert response.usage.total_tokens == 15
 
 
-def test_current_groq_sdk_usage_is_not_recognized() -> None:
+def test_known_limitation_groq_sdk_usage_is_not_recognized() -> None:
     groq = pytest.importorskip("groq.types.chat")
     response = groq.ChatCompletion.model_validate(
         {
@@ -216,14 +134,16 @@ def usage_chunk() -> ChatCompletionChunk:
     )
 
 
-def test_current_stream_extractor_discards_usage_only_chunk() -> None:
+def test_known_limitation_stream_extractor_discards_usage_only_chunk() -> None:
     handler = OpenAIJSONHandler()
     assert list(handler.extract_streaming_json([usage_chunk()])) == []
 
 
 @pytest.mark.asyncio
-async def test_current_async_stream_extractor_discards_usage_only_chunk() -> None:
-    async def chunks() -> AsyncIterator[ChatCompletionChunk]:
+async def test_known_limitation_async_stream_extractor_discards_usage_only_chunk() -> (
+    None
+):
+    async def chunks() -> AsyncGenerator[ChatCompletionChunk, None]:
         yield usage_chunk()
 
     handler = OpenAIJSONHandler()
@@ -232,7 +152,7 @@ async def test_current_async_stream_extractor_discards_usage_only_chunk() -> Non
     ] == []
 
 
-def test_current_batch_success_drops_per_item_usage() -> None:
+def test_known_limitation_batch_success_drops_per_item_usage() -> None:
     class Item(BaseModel):
         value: int
 
@@ -251,12 +171,13 @@ def test_current_batch_success_drops_per_item_usage() -> None:
         },
     }
     result = processor.parse_results(json.dumps(raw))[0]
+    assert isinstance(result, BatchSuccess)
     assert result.success
     assert set(result.model_dump()) == {"custom_id", "result", "success"}
     assert not hasattr(result.result, "_raw_response")
 
 
-def test_current_anthropic_beta_usage_is_not_recognized() -> None:
+def test_known_limitation_anthropic_beta_usage_is_not_recognized() -> None:
     beta = pytest.importorskip("anthropic.types.beta")
     response = beta.BetaMessage.model_validate(
         {
@@ -273,7 +194,9 @@ def test_current_anthropic_beta_usage_is_not_recognized() -> None:
     assert not has_compatible_usage(response, _initialize_usage(Provider.ANTHROPIC))
 
 
-def test_current_local_cache_replays_historical_raw_usage_without_total() -> None:
+def test_known_limitation_local_cache_replays_historical_raw_usage_without_total() -> (
+    None
+):
     from instructor.cache import AutoCache, load_cached_response, store_cached_response
     from instructor.v2.core.retry import _finalize_parsed_response
 
@@ -292,47 +215,15 @@ def test_current_local_cache_replays_historical_raw_usage_without_total() -> Non
     assert not hasattr(restored, "_total_usage")
 
 
-@pytest.mark.parametrize(
-    ("previous", "current", "expected"),
-    [(0, 0, 0), (None, 6, 6), (12, None, None), (True, 6, 6), (12, False, False)],
-)
-def test_anthropic_dictionary_thinking_preserves_non_counts(
-    previous: int | None, current: int | None, expected: int | None
-) -> None:
-    types = pytest.importorskip("anthropic.types")
-    if "output_tokens_details" in types.Usage.model_fields:
-        pytest.skip(
-            "Tests extra-dictionary behavior in SDKs without this declared field"
-        )
-    from instructor.v2.providers.anthropic.usage import initialize_usage
-    from instructor.v2.providers.anthropic.usage import update_total_usage as accumulate
-
-    total = initialize_usage()
-    for count, unknown in [(previous, 100), (current, 200)]:
-        usage = types.Usage.model_validate(
-            {
-                "input_tokens": 1,
-                "output_tokens": 20,
-                "output_tokens_details": {
-                    "thinking_tokens": count,
-                    "unknown_metadata": unknown,
-                },
-            }
-        )
-        assert accumulate(usage, total)
-    details = total.model_dump()["output_tokens_details"]
-    assert details["thinking_tokens"] == expected
-    assert type(details["thinking_tokens"]) is type(expected)
-    assert details["unknown_metadata"] == 200
-
-
-def test_current_cli_mini_model_uses_full_model_price_table() -> None:
+def test_known_limitation_cli_mini_model_uses_full_model_price_table() -> None:
     from instructor.cli.usage import get_model_cost
 
-    assert get_model_cost("gpt-4o-mini") is get_model_cost("gpt-4o")
+    # Deliberately exercise a runtime model name omitted from the legacy literal.
+    mini_cost = get_model_cost("gpt-4o-mini")  # ty: ignore[invalid-argument-type]
+    assert mini_cost is get_model_cost("gpt-4o")
 
 
-def test_current_usage_event_copy_is_shared_between_handlers() -> None:
+def test_known_limitation_usage_event_copy_is_shared_between_handlers() -> None:
     from instructor.v2.core.hooks import Hooks
     from instructor.v2.core.retry import _usage_snapshot
